# By Jonathan Chiam
# Via Jonathan Chiam (3) and Leon Chin (2)
* 'develop' of https://github.com/leonchinzeye/CS3201:
  Fixed all results class errors and misc build warnings
  Updated errors in Uses System Test with gitignore updates
  Fixed build error and removed unnecessary results files
  Modifies Clause Tests Done

Conflicts:
	SPA/WithClause.h
	UnitTesting/UnitTesting.vcxproj

+++ SPA/ProgramCNode.cpp
@@ -19,7 +19,7 @@ vector<CNode*>& ProgramCNode::getProcedures() {
 }
 
 bool ProgramCNode::contains(const string& procName) {
-	for(int i = 0; i < this->children.size(); i++) {
+	for(size_t i = 0; i < this->children.size(); i++) {
 		CNode* child = children.at(i);
 		if(child->getNodeType() == PROC_ && child->getName() == procName) {
 			return true;
@@ -30,7 +30,7 @@ bool ProgramCNode::contains(const string& procName) {
 }
 
 ProcCNode* ProgramCNode::getProcNode(const string& procName) {
-	for(int i = 0; i < this->children.size(); i++) {
+	for(size_t i = 0; i < this->children.size(); i++) {
 		CNode* child = children.at(i);
 		if(child->getNodeType() == PROC_ && child->getName() == procName) {
 			return (ProcCNode*) child;
